#include <stdlib.h> for putenv, <string.h> for strlen.
Wed Jan  8 18:44:21 2003  Manish Singh  <yosh@gimp.org>

        * gdk/x11/gdkdisplay-x11.c: #include <stdlib.h> for putenv,
        <string.h> for strlen.

        * gdk/x11/gdkscreen-x11.c: #include <stdlib.h> for getenv,
        <string.h> for strrchr.

        * tests/testgtk.c (create_saved_position): "x" and "y" properties
        for widgets no longer exist, use gtk_window_move instead.

        * test/testtextbuffer.c (check_get_set_text): cast strlen to
        int for g_error.

+++ tests/testtextbuffer.c
@@ -169,7 +169,7 @@ check_get_set_text (GtkTextBuffer *buffer,
 
   if (n != strlen (str))
     g_error ("Sum of chars in lines is %d but buffer byte count is %d",
-             n, strlen (str));
+             n, (int) strlen (str));
   
   gtk_text_buffer_set_text (buffer, "", -1);
